refactor(deps): update dependencies

+++ src/routes/admin/dashboard/users/index.tsx
@@ -6,7 +6,7 @@ import {
   z,
   zod$,
 } from "@builder.io/qwik-city";
-import { asc, desc, getTableColumns, like, or, sql } from "drizzle-orm";
+import { asc, desc, eq, getTableColumns, like, or, sql } from "drizzle-orm";
 import LucideTrash from "~/components/icons/lucide-trash";
 import { Button } from "~/components/ui/buttons";
 import type { CrudCookies } from "~/components/ui/crud";
@@ -34,7 +34,7 @@ export const useUsersCrudCookies = routeLoader$(async (event) => {
     event.cookie.set(
       Router.admin.dashboard.users.index,
       defaultCrudCookies,
-      CrudCookiesOptions
+      CrudCookiesOptions,
     );
   }
   const response = event.cookie
@@ -45,23 +45,23 @@ export const useUsersCrudCookies = routeLoader$(async (event) => {
 
 export const useCrudUsers = routeLoader$(async (event) => {
   const crudCookies = await event.resolveValue(useUsersCrudCookies);
-  let itemsQuery = await db
+  let itemsQuery = db
     .select()
     .from(users)
     .limit(crudCookies.limit)
     .offset(crudCookies.offset);
-  let countQuery = await db
+  let countQuery = db
     .select({ count: sql<number>`count(id)`.mapWith(Number) })
     .from(users);
 
   if (crudCookies.orderBy) {
     const columns = getTableColumns(users);
     const columnName = crudCookies.orderBy.split(
-      ","
+      ",",
     )[0] as keyof typeof columns;
     const sort = crudCookies.orderBy.split(",")[1] as "asc" | "desc";
     itemsQuery = itemsQuery.orderBy(
-      sort === "asc" ? asc(columns[columnName]) : desc(columns[columnName])
+      sort === "asc" ? asc(columns[columnName]) : desc(columns[columnName]),
     );
   }
 
@@ -69,7 +69,7 @@ export const useCrudUsers = routeLoader$(async (event) => {
     const query = or(
       like(users.id, `%${crudCookies.search}%`),
       like(users.email, `%${crudCookies.search}%`),
-      like(users.name, `%${crudCookies.search}%`)
+      like(users.name, `%${crudCookies.search}%`),
     );
     itemsQuery = itemsQuery.where(query);
     countQuery = countQuery.where(query);
@@ -77,7 +77,7 @@ export const useCrudUsers = routeLoader$(async (event) => {
 
   return {
     items: await itemsQuery.all(),
-    count: await countQuery.get().count,
+    count: (await countQuery.get()?.count) || 0,
     crudCookies,
   };
 });
@@ -90,11 +90,20 @@ export const useDeleteUser = routeAction$(
       withToast(event, ToastType.error, "You can not delete yourself!");
       return event.fail(419, {});
     }
+    const admin = await db
+      .select()
+      .from(users)
+      .where(eq(users.email, process.env.ADMIN_USER!))
+      .get();
+    if (admin?.id === input.id) {
+      withToast(event, ToastType.error, "You can not delete an admin user!");
+      return event.fail(419, {});
+    }
     await auth.deleteUser(input.id);
   },
   zod$({
     id: z.string(),
-  })
+  }),
 );
 
 // TODO: Add filtering
